refactor: room 관련 폴더 구조 변경
