+++ frontend/src/features/room/lobby/components/ParticipantSection/ParticipantSection.styled.ts
