+++ frontend/src/features/room/lobby/components/ParticipantSection/ParticipantSection.tsx
